backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
psycopg2-binary
redis==5.0.1
celery==5.3.4
django-celery-beat
transformers==4.37.2
torch==2.2.0
tensorflow==2.15.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.1
django-extensions==3.2.3
Pillow==10.2.0
dj-database-url